+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add handler para cubrir el login del admin
 - Add aula connect logo en el admin dashboard
 - Add patch edit users courses in admin dashboard allows duplicated course
+- Remove temporally the change view mode
+- Remove the pricing option
 
 **v1.1.0**